requirements.txt
# listado de librerias
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary

